XFL2SESX Part 2
What's New Scooby Doo?

Duration seems to be computed correctly, but the duration of the clips are too long. I don't think this is a logical error, but the sample count used to calculate duration might not be right.

Audition seems to patch out the <files> when you open the file. Audio Clip ids get set to our old friend, 2^32. I wrestled with this for a bit and it seems unavoidable. A workaround exists, however.

This conversion process only accounts for the bare minimum requirements to construct the file and Audition fills in the gaps. The first time opening the file steamrolls our existing <files> structure as well as the Audio Clip ids. However, changing them after the first save does not reset them. In this case, we may have three steps, one where we generate the file, two, we open the file, three, we patch the file. Patching the file would consist of reading the XML structure and binding the AudioClip ids to a reconstructed <file> structure that we append after document. This is shitty but it would work.

Can probably do all this in one more day.