Fix longstanding bug where strokes would not be converted to SVG paths properly
The main cause of this issue was that in obtaining all of the pointLists
for each fillStyle and strokeStyle of an XFL Edges in ConvertEdgesToSvgPath() in EdgeUtils,
only copies of the references to these lists were being added. In most cases,
this is fine. However this can cause a problem if the same pointList
appears for both a fillStyle and a strokeStyle.

When processing pointLists for fillStyles, they are first converted into shapes
before being converted to SVG path strings. In contrast, for strokeStyles,
they are just converted outright. If a fillStyle and strokeStyle were to
both have the same pointList stored, in processing the former, the pointList
would be modified for both fillStyle and strokeStyle (as the reference to
the same list is stored in both).

To prevent pointLists for changing for the strokeStyle, I created a deep
copy of the list before adding it to the strokeStyle's pointList list.
While this can introduce a lot more space depending on how many pointLists
are associated with a strokeStyle in a XFL Edges, it gurantees that the strokeStyle
will have an indepedent copy of that pointList.

This can be potentially improved upon in the future by using a HashSet to keep
track of the same pointLists in both the fillStyles and strokeStyles so that
only a deep copy would be created for the pointLists that need it.